Correcoes novamente

+++ src/main/java/com/dssystems/dssystemsbackend/models/Transacao.java
@@ -32,7 +32,7 @@ public class Transacao {
     public Transacao() {}
 
     public Integer getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    public void setId(Integer id) { this.id = id; }
 
     public Usuario getUsuario() { return usuario; }
     public void setUsuario(Usuario usuario) { this.usuario = usuario; }
